perf(api-tests): boot test PGlite from a per-run schema snapshot and split RSS budgets
Every DB-touching test process used to run drizzle-kit's pushSchema
itself, paying a ~2.2 GB peak that forced a single 4096 MB batch
budget sized for that floor. The batching runner now builds the schema
once per run (build-pglite-snapshot.ts), dumps the PGlite data dir,
and test processes boot from the snapshot via loadDataDir without
loading drizzle-kit at all; solo bun-test runs keep the full in-process
build as a fallback. The peak-RSS guard now carries separate budgets:
2560 MB for DB batches (worst measured snapshot-booted batch 2072 MB)
and 2048 MB for pure-logic batches (worst measured 1520 MB), so a real
memory regression trips at a meaningful size instead of hiding under a
4 GB ceiling.
