+++ apps/api/src/lib/entity-filters.differential.test.ts
@@ -1,6 +1,5 @@
 import { afterAll, beforeAll, expect, test } from "bun:test";
-import { pushSchema } from "drizzle-kit/api-postgres";
-import { and, eq, sql } from "drizzle-orm";
+import { and, eq } from "drizzle-orm";
 import { drizzle } from "drizzle-orm/pglite";
 import fc from "fast-check";
 
@@ -8,16 +7,12 @@ import type { ConditionNode } from "@stll/conditions";
 import { propertyConfig, propertyTestTimeout } from "@stll/property-testing";
 
 import * as authSchema from "@/api/db/auth-schema";
-import * as rlsExports from "@/api/db/rls";
 import * as schema from "@/api/db/schema";
 import { entities, entityVersions, fields, properties } from "@/api/db/schema";
 import type { EntityKind, FieldContent } from "@/api/db/schema-validators";
 import { toSafeId } from "@/api/lib/branded-types";
 import { applyFilters, buildFilterConditions } from "@/api/lib/entity-filters";
-import {
-  createSchemaPglite,
-  installPgliteSchemaPrerequisites,
-} from "@/api/tests/pglite-schema";
+import { createTestPglite } from "@/api/tests/pglite-test-db";
 
 // ── Differential property test ──────────────────────────────
 //
@@ -29,11 +24,9 @@ import {
 // a real bug in one of the two implementations (or a by-design divergence
 // that is deliberately excluded from the generator below, with a reason).
 
-const allSchema = { ...schema, ...authSchema, ...rlsExports };
-
 type RawDb = ReturnType<typeof drizzle>;
 
-let client: Awaited<ReturnType<typeof createSchemaPglite>> | undefined;
+let client: Awaited<ReturnType<typeof createTestPglite>> | undefined;
 let db: RawDb;
 
 const ORG_ID = toSafeId<"organization">(Bun.randomUUIDv7());
@@ -95,20 +88,8 @@ const ENTITY_KINDS: readonly EntityKind[] = [
 
 beforeAll(
   async () => {
-    client = await createSchemaPglite();
+    client = await createTestPglite();
     db = drizzle({ client });
-    const pushDb = drizzle({ client });
-
-    await db.execute(sql.raw("CREATE ROLE stella NOLOGIN"));
-    await db.execute(sql.raw("CREATE ROLE stella_ingestion NOLOGIN"));
-    await installPgliteSchemaPrerequisites(db);
-
-    const { sqlStatements } = await pushSchema(allSchema, pushDb);
-    // Schema DDL must apply in dependency order, so these run sequentially.
-    for (const statement of sqlStatements) {
-      // eslint-disable-next-line no-await-in-loop -- ordered DDL, can't parallelize
-      await db.execute(sql.raw(statement));
-    }
 
     await db.insert(authSchema.user).values({
       id: "user_diff_test",
